move filtered menu items into separate component
