+++ client/src/components/Header.tsx
@@ -1,0 +1,7 @@
+export default function Header() {
+    return (
+        <div className="text-left py-4 bg-white shadow-md">
+            <div className=" ml-2 text-xl font-bold text-gray-800"><h1>YES CHEF</h1></div>
+        </div>
+    )
+}
